+++ src/main/java/com/github/nthykier/debpkg/deb822/dialects/Deb822DialectDebianTestsControlLanguage.java
@@ -28,4 +28,9 @@ public class Deb822DialectDebianTestsControlLanguage extends Language implements
         return DEBIAN_TESTS_CONTROL_PARAGRAPH_CLASSIFIER;
     }
 
+    @Override
+    public boolean supportsEmptyFields() {
+        return true;
+    }
+
 }